refactor: move table files to sepearate files
