+++ chap_core/database/model_template_tables.py
@@ -1,0 +1,43 @@
+from typing import Optional, List
+
+from sqlalchemy import Column, JSON
+from sqlmodel import SQLModel, Field, Relationship
+
+from chap_core.database.base_tables import DBModel
+from chap_core.model_spec import PeriodType
+
+
+class ModelTemplateMetaData(SQLModel):
+    display_name: str = 'No Display Name yet'
+    description: str = "No Description yet"
+    author: str = "Unknown Author"
+    organization: Optional[str] = None
+    organization_logo_url: Optional[str] = None
+    contact_email: Optional[str] = None
+    citation_info: Optional[str] = None
+
+
+class ModelTemplateInformation(SQLModel):
+    supported_period_type: PeriodType = PeriodType.any
+    user_options: Optional[dict] = Field(default_factory=dict, sa_column=Column(JSON))
+    required_covariates: List[str] = Field(default_factory=list, sa_column=Column(JSON))
+    target: str = 'disease_cases'
+    allow_free_additional_continuous_covariates: bool = False
+
+
+class ModelTemplateDB(DBModel, ModelTemplateMetaData, ModelTemplateInformation, table=True):
+    '''
+    TODO: Maybe remove Spec from name, or find common convention for all models.
+    Just a mixin here to get the model info flat in the database.
+    '''
+    name: str
+    id: Optional[int] = Field(primary_key=True, default=None)
+    source_url: Optional[str] = None
+
+
+class ConfiguredModelDB(DBModel, table=True):
+    name: str
+    id: Optional[int] = Field(primary_key=True, default=None)
+    model_template_id: int = Field(foreign_key="modeltemplatedb.id")
+    model_template: ModelTemplateDB = Relationship()
+    configuration: Optional[dict] = Field(sa_column=Column(JSON))
