+++ chap_core/file_io/file_paths.py
@@ -6,6 +6,10 @@ def get_results_path():
     path.mkdir(exist_ok=True)
     return path
 
+def get_example_data_path():
+    path = Path(__file__).parent.parent.parent / "example_data"
+    return path
+
 
 def get_models_path():
     path = Path(__file__).parent.parent.parent / "external_models"
